Create calibration scan version of Excitation729
This scan is indentical to Excitation729, except that it uses a few key parameters which are independent of those used for Excitation729 (specifically, they are in the collection Excitation_729_Cal1). This allows one to schedule this scan as a calibration, and use the same parameters each time the calibration occurs, independent of how other stuff is changed in the meantime. Currently intend to use it to keep track of the 866 frequency. Scan name ends in the number 1 in anticipation of the possibility of the scan being copied again to allow multiple parallel independent running calibrations.